mqtt-ble-radiator-valve/pythonProject/requirements.txt
aioesphomeapi==27.0.0
pyyaml
aiomqtt